feat(demo5): configurable cycle pause, 32-bit stats counters, stats on by default
Make the inter-measurement pause configurable via DS18B20_CYCLE_PAUSE_US
(default 5s to preserve existing behaviour; demo5 overrides to ~1ms) instead
of the hard-coded 5s PAUSE_5S.  Widen every statistics counter and the
histogram bucket array from uint16_t to uint32_t so a long accumulation window
(no longer overflows past 65535 samples) — this is required now that the stats
period can span thousands of rounds.

demo5 now builds with OW_STATS_ENABLE, STATS_DUMP_INTERVAL=5000 and a 1ms
cycle pause by default.  Parasite power is deliberately left as an explicit
opt-in (EXT="-DPARASITE_POWER=1") because it is bus-hardware dependent, not a
property of the demo.  Document the extra flag in demo5.c.

Add a unit test proving the counters survive >65535 ticks/samples without
wrapping.  Update the test accessor shims and assertions to uint32_t.

+++ inc/ow_stats.h
@@ -34,10 +34,10 @@ typedef struct {
     uint8_t rom[8]; /**< 64-bit ROM address (LSB first) */
     uint8_t min_pulse; /**< Shortest observed pulse width (us) */
     uint8_t max_pulse; /**< Longest  observed pulse width (us) */
-    uint16_t count; /**< Successful scratchpad reads */
-    uint16_t crc_err; /**< CRC-8 mismatches */
-    uint16_t no_presence; /**< No presence pulse detected */
-    uint16_t generic_err; /**< Unexpected / reserved-byte errors */
+    uint32_t count; /**< Successful scratchpad reads */
+    uint32_t crc_err; /**< CRC-8 mismatches */
+    uint32_t no_presence; /**< No presence pulse detected */
+    uint32_t generic_err; /**< Unexpected / reserved-byte errors */
 } ow_stats_sensor_t;
 
 /**
@@ -46,9 +46,9 @@ typedef struct {
 typedef struct {
     ow_stats_sensor_t sensors[OW_STATS_MAX_SENSORS];
     uint8_t sensor_count; /**< Number of known sensors */
-    uint16_t histogram[OW_STATS_HIST_BUCKETS]; /**< Pulse-width distribution */
-    uint16_t total_cycles; /**< Full measurement cycles completed */
-    uint16_t total_errors; /**< Sum of all error counters */
+    uint32_t histogram[OW_STATS_HIST_BUCKETS]; /**< Pulse-width distribution */
+    uint32_t total_cycles; /**< Full measurement cycles completed */
+    uint32_t total_errors; /**< Sum of all error counters */
 } ow_stats_t;
 
 /**
@@ -103,7 +103,7 @@ void ow_stats_reset(void);
  * @brief Increment the cycle counter (called from the demo callback).
  * @return The new cycle count.
  */
-uint16_t ow_stats_tick(void);
+uint32_t ow_stats_tick(void);
 
 #else /* OW_STATS_ENABLE not defined — zero-overhead stubs */
 
@@ -121,7 +121,7 @@ static inline void ow_stats_count_error(int16_t error, const uint8_t* rom) {
 static inline void ow_stats_dump_start(void) {}
 static inline uint8_t ow_stats_dump_poll(void) { return 1; }
 static inline void ow_stats_reset(void) {}
-static inline uint16_t ow_stats_tick(void) { return 0; }
+static inline uint32_t ow_stats_tick(void) { return 0; }
 
 #endif /* OW_STATS_ENABLE */
